fix(day9-2): correct tail behavior
I misunderstood the original instructions for how the tail should follow
the head. It's a wonder that I got the first part and the tests passing
that way lol.

The problem was that I wasn't really handling the case where, if the
tail and head aren't touching and they aren't in the same rank/file, the
tail should move diagonally towards the head until they touch.